✂️📙 Checkpoint ./src/presenters/pop-overs/add-team-user-pop.js:271885/178 ./src/presenters/includes/team-users.js:271885/3067 ./src/presenters/pop-overs/team-user-info-pop.js:271885/1399 ./src/presenters/pop-overs/add-project-to-collection-pop.js:271885/917 ./src/presenters/includes/add-project-to-collection.js:271885/684 ./src/presenters/pop-overs/project-options-pop.js:271885/62 ./src/components/project/project-embed.js:271885/20
