+++ src/components/project/project-embed.js
@@ -6,7 +6,7 @@ import Embed from 'Components/project/embed';
 
 import ReportButton from '../../presenters/pop-overs/report-abuse-pop';
 import { EditButton, RemixButton } from '../../presenters/includes/project-actions';
-import AddProjectToCollection from '../../presenters/includes/add-project-to-collection';
+import AddProjectToCollection from '../../presenters/pop-overs/add-project-to-collection';
 import { useTracker } from '../../presenters/segment-analytics';
 
 import styles from './project-embed.styl';
