+++ src/presenters/pop-overs/project-options-pop.js
@@ -6,7 +6,7 @@ import PopoverWithButton from './popover-with-button';
 import { NestedPopover } from './popover-nested';
 import { useCurrentUser } from '../../state/current-user';
 
-import AddProjectToCollectionPop from './add-project-to-collection-pop';
+import { AddProjectToCollectionBase } from './add-project-to-collection-pop';
 
 const isTeamProject = ({ currentUser, project }) => {
   for (const team of currentUser.teams) {
@@ -226,7 +226,7 @@ export default function ProjectOptionsPop(props) {
       {({ togglePopover, focusFirstElement }) => (
         <NestedPopover
           alternateContent={() => (
-            <AddProjectToCollectionPop
+            <AddProjectToCollectionBase
               {...props}
               {...projectOptions}
               togglePopover={togglePopover}
